% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric_comma.r
\name{as_numeric_comma}
\alias{as_numeric_comma}
\title{Converts Character to Numeric for Numbers Using a Decimal Comma}
\usage{
as_numeric_comma(.x, .group.separator = ".")
}
\arguments{
\item{.x}{Character vector. The first "," of each entry is turned into a "."
and \code{as.numeric()} is applied to the result.}

\item{.group.separator}{Character value. This character is removed from the
input before conversion. Useful to remove "." used
as grouping character (e.g. "1.234,56")}
}
\value{
A numeric vector of same length as \code{.x}.
}
\description{
This function turns character variables containing numbers but using a
decimal comma, e.g. "1,23" into a numeric.
}
\examples{
as_numeric_comma(c("1,234", "4", "1.234", "1,2,3"))

# .group.separator:
as_numeric_comma(c("1.234,56", "1.234"))
as_numeric_comma(c("1.234,56", "1.234"), .group.separator = ".")

}
