% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bartext.r
\docType{data}
\name{geom_bartext}
\alias{geom_bartext}
\alias{GeomBarText}
\title{Bar Plot with Numbers at the Ends of the Bars. Options not mentioned here
work identical to \code{geom_col} or \code{geom_text}.}
\format{An object of class \code{GeomBarText} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
geom_bartext(mapping = NULL, data = NULL, position = "identity", ...,
  width = NULL, just_mirror_cutoff = 0.9, fontcolor = c("grey15",
  "grey85"), parse = FALSE, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

GeomBarText
}
\arguments{
\item{mapping}{\itemize{
\item 
}}

\item{data}{\itemize{
\item 
}}

\item{position}{\itemize{
\item 
}}

\item{...}{\itemize{
\item 
}}

\item{width}{\itemize{
\item 
}}

\item{just_mirror_cutoff}{Numeric value between 0 and 1. The default of 0.9
means that if the text is supposed to appear
\emph{outside} the bars then for values larger (in
absolute terms) than 90% of the data this is
flipped to appear inside. And vice versa (in that
case a value of e.g. 0.1 will make more sense).
This avoids text getting cutoff by the edge of
the plot or text obscuring the bars for very
small values. Set it to 0 or 1 to switch off.}

\item{fontcolor}{Character vector of color specifications (names, hex-values,
etc.). For text appearing inside the bars the geom will
use the value "most unlike" the bar color. There are some
limits to this and bugs to be found.}

\item{parse}{\itemize{
\item 
}}

\item{na.rm}{\itemize{
\item 
}}

\item{show.legend}{\itemize{
\item 
}}

\item{inherit.aes}{\itemize{
\item 
}

Relevant aes are
\itemize{
\item \strong{x}
\item \strong{y}
\item label - \code{y} is used otherwise
}

The following \code{aes} are passed to \code{geom_col}
\itemize{
\item color
\item fill
\item size
\item linetype
\item alpha
}

The following \code{aes} are used by \code{geom_text}
\itemize{
\item fontsize - the size \code{aes} of geom_text
\item angle
\item \strong{injust} - like \code{hjust} and \code{vjust} but it represents the "inward"
justification. A value of 0 will be just outside and a
value of 1 just inside the bar border.
\item family
\item fontface
\item lineheight
}}
}
\description{
Bar Plot with Numbers at the Ends of the Bars. Options not mentioned here
work identical to \code{geom_col} or \code{geom_text}.
}
\examples{
library(ggplot2)

X <- data.frame(x = 1:5, y = sample(1:100, 5))

ggplot(X, aes(x, y)) +
  geom_bartext(injust = -0.5)

ggplot(X, aes(x, y)) +
  geom_bartext(injust = 1.5)

ggplot(X, aes(x, y)) +
  geom_bartext(injust = 1.5) +
  coord_flip()

ggplot(X, aes(x, y, label = percentify(y))) +
  geom_bartext(injust = -0.5) +
  coord_flip()

}
\keyword{datasets}
