% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_coltext.r
\docType{data}
\name{geom_coltext}
\alias{geom_coltext}
\alias{geom_bartext}
\alias{GeomColtext}
\title{Bar Plot with Numbers at the Ends of the Bars. Options not mentioned here
work identical to \code{geom_col} or \code{geom_text}.}
\format{An object of class \code{GeomColtext} (inherits from \code{GeomRect}, \code{Geom}, \code{ggproto}) of length 5.}
\usage{
geom_coltext(mapping = NULL, data = NULL, position = "identity", ...,
  stat = "identity", width = NULL, center_text = FALSE,
  just_mirror = TRUE, just_mirror_cutoff = c(0.1, 0.9),
  fontcolor = c("grey15", "grey85"), parse = FALSE, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

geom_bartext(mapping = NULL, data = NULL, position = "identity", ...,
  stat = "identity", width = NULL, center_text = FALSE,
  just_mirror = TRUE, just_mirror_cutoff = c(0.1, 0.9),
  fontcolor = c("grey15", "grey85"), parse = FALSE, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

GeomColtext
}
\arguments{
\item{mapping}{\itemize{
\item 
}}

\item{data}{\itemize{
\item 
}}

\item{position}{\itemize{
\item 
}}

\item{...}{\itemize{
\item 
}}

\item{width}{\itemize{
\item 
}}

\item{center_text}{Logical value. If \code{TRUE} the text will be centered within
its respective bar.}

\item{just_mirror}{Logical value. If \code{TRUE} then the justification of the
text will be mirrored at the edge of the bar for the
largest and smallest values (see \code{just_mirror_cutoff}
below). E.g. if the text is outside the bars then for
the largest values it will be inside the bars.
This prevents text from being cut off by the
edges of the plot.}

\item{just_mirror_cutoff}{Two numeric values between 0 and 1. The default of
0.9 for the second value
means that if the text is supposed to appear
\emph{outside} the bars then for values larger (in
absolute terms) than 90% of the data range this is
flipped to appear inside. And vice versa for text
inside bars and the first value of just_mirror_cutoff.
This avoids text getting cutoff by the edge of
the plot or text obscuring the bars for very
small values. Set it to 0 or 1 to switch off.}

\item{fontcolor}{Character vector of color specifications (names, hex-values,
etc.). For text appearing inside the bars the geom will
use the value "most unlike" the bar color. There are some
limits to this and bugs to be found.}

\item{parse}{\itemize{
\item 
}}

\item{na.rm}{\itemize{
\item 
}}

\item{show.legend}{\itemize{
\item 
}}

\item{inherit.aes}{\itemize{
\item 
}

Relevant aes are
\itemize{
\item \strong{x}
\item \strong{y}
\item label - \code{y} is used otherwise
}

The following \code{aes} are passed to \code{geom_col}
\itemize{
\item color
\item fill
\item size
\item linetype
\item alpha
}

The following \code{aes} are used by \code{geom_text}
\itemize{
\item fontsize - the size \code{aes} of geom_text
\item angle
\item \strong{injust} - like \code{hjust} and \code{vjust} but it represents the "inward"
justification. A value of -1 will be just outside and a
value of 1 just inside the bar border.
\item family
\item fontface
\item lineheight
}}
}
\description{
Bar Plot with Numbers at the Ends of the Bars. Options not mentioned here
work identical to \code{geom_col} or \code{geom_text}.
}
\examples{
library(ggplot2)

X <- data.frame(x = 1:5, y = sample(1:100, 5))

ggplot(X, aes(x, y)) +
  geom_bartext(injust = -0.5)

ggplot(X, aes(x, y)) +
  geom_bartext(injust = 1.5)

ggplot(X, aes(x, y)) +
  geom_bartext(injust = 1.5) +
  coord_flip()

ggplot(X, aes(x, y, label = percentify(y))) +
  geom_bartext(injust = -0.5) +
  coord_flip()

}
\keyword{datasets}
