% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brightness.r
\name{get_brightness}
\alias{get_brightness}
\title{Get Brightness of a Color}
\usage{
get_brightness(.color)
}
\arguments{
\item{.colors}{Vector or list with colors.This is passed to \code{col2rgb} and
so accepts color names ("black"), hexadecimal rgb strings
("#rrggbb" or "#rrggbbaa") or positive integers (\code{i} meaning
\code{palette()[i]}).}
}
\value{
A vector of brightness values between 0 and 1. 0 meaning black and
1 white.
}
\description{
This function tries to return how "bright" (i.e. perceived luminance) a given
color is based on psychological evaluations at the dawn of the color TV era.
}
\details{
This takes into account that the human eye is most sensitive to green,
followed by red with blue colors appearing darker.

Everything is stolen from stackoverflow so don't sue me.
}
