% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\alias{notin}
\title{Negation of \%in\% Operator}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{Vector. The values to be matched}

\item{table}{Vector. The values to be matched against.}
}
\value{
A logical vector of same length as \code{x}. \code{TRUE} means the respective
value was found in \code{table}.
}
\description{
R lacks a negation of the \%in\% operator. While you can negate the whole
expression, that solution results in less readable and clear code.
}
\examples{
c("a", "ab") \%notin\% letters

}
