% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_coltext.r
\name{\%||\%}
\alias{\%||\%}
\title{Import of ggplot NULL-Avoiding OR}
\usage{
a \%||\% b
}
\arguments{
\item{a}{LHS. The operator returns this value unless it is \code{NULL}.}

\item{b}{RHS. If the LHS is \code{NULL} then this is return}
}
\value{
Either LHS or RHS
}
\description{
Import of ggplot NULL-Avoiding OR
}
