% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stri_s.R
\name{\%s\%}
\alias{\%s\%}
\title{Export of stringi string concatenate}
\usage{
e1 \%s\% e2
}
\arguments{
\item{e1}{Character vector. LHS}

\item{e2}{Chracter vector. RHS

The operator concatenates the LHS and RHS just as if you'd typed
\code{stri_c(e1, e2)} or \code{str_c(e1, e2)} respectively.}
}
\description{
Export of stringi string concatenate
}
\examples{
letters \%s\% LETTERS

}
