% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_dropdown.r
\name{insert_dropdown_html}
\alias{insert_dropdown_html}
\alias{insert_dropdown_location}
\title{Insert the Necessary Code to Turn Tabs into Dropdowns}
\usage{
insert_dropdown_html()

insert_dropdown_location()
}
\description{
RMarkdown allows you to turn subsections into tabs by putting \code{{.tabset}}
behind the title of the parent section. With this code you can write
\code{{.tabset .tabset-dropdown}} to get a dropdown menus instead of tabs.
}
\details{
To use this, either create a chunk (with \code{results = "asis"}) that calls
\code{insert_dropdown_html}, or put the value of \code{insert_dropdown_location()} as
\code{after_body} include in the RMarkdown.

\preformatted{---
title: Dropdowns
output: html_document
---

# Header1 {.tabset .tabset-dropdown}

## Dropdown Entry 1
...

## Dropdown Entry 2
...

```{r, results = 'asis', echo = FALSE}
statupinternal::insert_dropdown_html()
```
}
}
