% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_dropdown.r
\name{insert_dropdown_location}
\alias{insert_dropdown_location}
\alias{insert_dropdown_html}
\title{Insert the Necessary Code to Turn Tabs into Dropdowns}
\usage{
insert_dropdown_location()

insert_dropdown_html()
}
\description{
RMarkdown allows you to turn subsections into tabs by putting \code{{.tabset}}
behind the title of the parent section. With this code you can write
\code{{.tabset .tabset-dropdown}} to get a dropdown menus instead of tabs.
}
\details{
To use this, either create a chunk (with \code{results = "asis"}) that calls
\code{insert_dropdown_html}, or put \code{insert_dropdown_location()} as \code{after_body}
include in the RMarkdown.

\preformatted{```{r, results = 'asis'}
insert_dropdown_html()
```}

or

\preformatted{---
title: Dropdowns
output:
  html_document:
    includes:
      after_body: `insert_dropdown_location()`
---

# Header1 {.tabset .tabset-dropdown}

## Dropdown Entry 1
...

## Dropdown Entry 2
...
}
}
