% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFALSE.r
\name{isFALSE}
\alias{isFALSE}
\title{Opposite to isTRUE}
\usage{
isFALSE(.x)
}
\arguments{
\item{.x}{Just about anything (as that just means that \code{isFALSE} is \code{FALSE})}
}
\description{
\code{isFALSE(.x)} is an abbreviation of \code{identical(FALSE, .x)}, and so is true if
and only if x is a length-one logical vector whose only element is \code{TRUE} and
which has no attributes (not even names).
}
\examples{
isFALSE(FALSE)
isFALSE(0)
isFALSE(c(FALSE, FALSE))

}
