% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_install.r
\name{load_or_install}
\alias{load_or_install}
\title{Install (If Necessary) and Load Multiple Packages at Once}
\usage{
load_or_install(..., .packages = NULL)
}
\arguments{
\item{...}{Packages. Quoted or unquoted.}

\item{.packages}{Character vector or list of packages. Mainly useful for
automated use in functions. Packages in .packages will be loaded after
those in ...}
}
\value{
TRUE if packages were installed, FALSE otherwise. There are no other
sanity checks at the moment.
}
\description{
Convenience function that for a list of packages looks if they are installed,
installs them if they are not and then loads them.
}
\examples{
load_or_install(dplyr, data.table, magrittr)

}
