% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outside.r
\name{outside}
\alias{outside}
\alias{\%outside\%}
\alias{\%notbetween\%}
\title{Negation of \%between\% Operator}
\usage{
outside(.x, .lower, .upper, .incbounds = TRUE)

.x \%outside\% .bounds

.x \%notbetween\% .bounds
}
\arguments{
\item{.x}{Numeric vector. Input data}

\item{.lower}{Numeric value or vector with the lower bound(s)}

\item{.upper}{Numeric value or vector with the upper bound(s)}

\item{.incbounds}{\code{TRUE} means inclusive bounds, i.e. \code{lower <= .x <= .upper}. \code{FALSE}
means exclusive bounds. For \%outside\%
\code{.incbounds = TRUE} is used.}
}
\value{
Logical vector of same length as \code{.x}. \code{TRUE} if the input value is
outside the bounds, \code{FALSE} otherwise.
}
\description{
R (data.table) lacks a negation of the \%between\% operator. While you can negate the whole
expression, that solution results in less readable and clear code.
}
\examples{
rnorm(10) \%outside\% c(-0.5, 0.5)

}
