% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentify.r
\name{percentify}
\alias{percentify}
\title{Turns a Probability into a Percentage}
\usage{
percentify(.x, .digits = 0)
}
\arguments{
\item{.x}{Numeric vector. Input data}

\item{.digits}{Integer value. To how many digits should the result be rounded.}
}
\value{
Vector with rounded percentages + "%" for use in graphical output.
}
\description{
The function turns a probability or share (e.g. 0.49, 2.102) into a percentage
(49%, 210.2%).
}
\examples{
percentify(1:10 / 10)
percentify(rnorm(10), 2)

}
