% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide.r
\name{provide}
\alias{provide}
\title{Restrict Variables to a Local Block}
\usage{
provide(..., .expr = NULL, .exports = NULL, .conditions = NULL,
  .env = parent.frame(), .quiet = FALSE)
}
\arguments{
\item{...}{Object definitions and the main expression.}

\item{.expr}{Expression to evaluate. By default the last element of ...}

\item{.exports}{List of objects to export. By default taken from everything but the last element of ...}

\item{.conditions}{List of conditions to check. Usually defined in ...}

\item{.env}{Environment. Where to provide the objects. By default the
environment where the function was executed.}

\item{.quiet}{Logical. Should the exports be printed to stdout?}
}
\description{
More to come
}
\examples{

# This provides X and Y. The variables a and beta are not exported
# as they are not specified in ...
provide(
  X = list(dim(X) == c(100,4),
           is.numeric(X)),
  Y,
  NROW(Y)==NROW(X),
  {
    a <- 4

    X <- matrix(rnorm(100*a), ncol=a)
    beta <- rnorm(a)

    Y <- X \%*\% beta + rnorm(100)
  }
)

}
