% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_sensibly.r
\name{round_sensibly}
\alias{round_sensibly}
\title{Round Numbers to an "Appropriate" Number of Digits}
\usage{
round_sensibly(.x, .digits = 0, .separation = TRUE)
}
\arguments{
\item{.x}{Numeric vector. Values to be rounded.}

\item{.digits}{Numeric vector. Number of digits that should be rounded to
relative to the base value decided by \code{.separation=}.}

\item{.separation}{Logical scalar. If \code{TRUE``}.digits=\code{is relative to the number of digits required to discern **all** values. If}FALSE` it's relative to the number of digits required to
discern the lowest and highest values.}
}
\value{
A rounded numeric vector.
}
\description{
This function is primarily geared towards use in labelling. By default it will
round a numeric vector to the lowest number of digits required to make sure
that no two values get rounded to the same value.
Alternatively it will round to significant digits, significant relative to
the \code{range} of the input vector.
}
\details{
The usage of \code{.digits} is not quite consistent between \code{.separation = TRUE}
and \code{.separation = FALSE}. But it was chosen this way because it corresponds
in my experience with the way we think about rounding numbers for labels.

Either you want "x significant digits" as in digits that make actually a
difference. It makes sense that 1 should be one significant digit and so on.

Or you want enough digits so that there are no identical labels. In this case
it makes sense that 0 should be that number of digits.
}
\examples{
x <- c(1.123, 1.134, 2.23)

# rounded to just enough digits to be able to separate one and 2
round_sensibly(x)

# rounded to one more digit
round_sensibly(x, 1)

# rounded to 1 significant digit of separation.
round_sensibly(x, 1, .separation = FALSE)

}
