% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_significantly.r
\name{round_significantly}
\alias{round_significantly}
\title{Round Numbers to Significant Digits}
\usage{
round_significantly(.x, .digits = 0)
}
\arguments{
\item{.x}{Numeric vector. Values to be rounded.}

\item{.digits}{Numeric vector. Number of significant digits you want to keep.}
}
\value{
A rounded numeric vector.
}
\description{
This function is primarily geared towards use in labelling. By default it will
round a numeric vector to a given number of significant digits. E.g. rounding
\code{0.00412} with \code{.digits = 1} will result in \code{0.004}, while \code{234} will result
in \code{200}.
}
\examples{
x <- c(-0.0002352, 1234)

round_significantly(x)
round_significantly(x, 2)

}
