% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageDocumentation.r
\docType{package}
\name{statupinternal-package}
\alias{statupinternal}
\alias{statupinternal-package}
\title{statupinternal: STAT-UP Utility Functions}
\description{
Miscallenous utility functions that haven't found a better place
yet. There is no other overarching theme to the package.
}
\keyword{internal}
