% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlike.r
\name{unlike}
\alias{unlike}
\alias{\%unlike\%}
\alias{\%notlike\%}
\title{Negation of \%like\% Operator}
\usage{
unlike(.vector, .pattern)

.vector \%unlike\% .pattern

.vector \%notlike\% .pattern
}
\arguments{
\item{.vector}{Either a character vector or a factor.}

\item{.pattern}{Gets passed on to \code{pattern} in \code{grepl}.}
}
\value{
Logical vector of same length as \code{.vector}. \code{TRUE} if the input value
matches the pattern (via regular expression matching). \code{FALSE} otherwise.
}
\description{
R (data.table) lacks a negation of the \%like\% operator. While you can negate the whole
expression, that solution results in less readable and clear code.
}
\examples{
"ana" \%like\% c("Anabelle", "banana", "BanAna", "meh")

}
