% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.r
\name{where}
\alias{where}
\title{Multiple ifelse clauses}
\usage{
where(.x, ...)
}
\arguments{
\item{.x}{Vector of values to match against}

\item{...}{Either formulas with a condition as LHS (optional) and an action
as RHS. Or named arguments that define additional values to match against,
i.e. that can be used in the formulas just like .x.}
}
\value{
The value resulting from the first valid condition \emph{for the given
vector element} is returned. Formulas without LHS or unnamed values are
used as default.
}
\description{
This is a version of ifelse that allows to check against multiple conditions
or alternatively a vectorized version of the \code{when} function in purrr.
}
\examples{
where(1:10,
      y = 10:1,
      .x <= 4 ~ y,
      .x <= 6 ~ 11:20,
      ~ -.x)
}
